Then God said, ‘Let there be hunger’; and there was hunger.
- The Empachado class is renamed at Stomach.
- Now the Stomach manages hunger. When it pass some defined time, it is
  added the hunger level to the actual level of the stomach. Defined by
a constant, the hunger level is negative, so it actually the
stomach actual level.
- Now dt is pass to the handleCollisions function. And at the same time,
  this function pass the dt to the update functions of the list of
objects to update. This is changed in case some of the updates depends
on the dt, so for convention is passed to everyone as the first
argument. In this case we need it for the stomach update, to calculate
the timer of the hunger.

(No se porqué decidí hacer las cosas en inglés, toy muriendo).